v2.0.0: Massive upgrade
