+++ requirements.txt
@@ -1,4 +1,2 @@
 Patrowl4py==1.0.3
-python-dateutil==2.8.1
-pytz==2019.3
 requests==2.22.0
